requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.1.3
Flask-HTTPAuth==4.7.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
outcome==1.2.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.1
